scsi: storvsc: Spread interrupts when picking a channel for I/O requests
Update the algorithm in storvsc_do_io to look for a channel starting
with the current CPU + 1 and wrap around (within the current NUMA
node). This spreads VMbus interrupts more evenly across CPUs. Previous
code always started with first CPU in the current NUMA node, skewing the
interrupt load to that CPU.

Signed-off-by: Michael Kelley <mikelley@microsoft.com>
Reviewed-by: Long Li <longli@microsoft.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>