drm/tegra: sor: Remove redundant definition
The TEGRA_IO_PAD_HDMI_DP0 definition was added in order to simplify the
dependency handling between the DRM and ARM-SoC trees. Since the change
from the ARM-SoC tree has now made it into v4.16-rc1, the definition is
no longer required because it is contained in include/soc/tegra/pmc.h.

Signed-off-by: Thierry Reding <treding@nvidia.com>